Test
Cac kieu

+++ src/main/java/com/msstore/DAO/BinhLuanDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.msstore.Entity.BinhLuan;
 
-public interface BinhLuanDAO extends JpaRepository<BinhLuan, Integer>{
+public interface BinhLuanDAO extends JpaRepository<BinhLuan, Long>{
 
 }
