+++ src/main/java/com/msstore/DAO/DonHangDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.msstore.Entity.DonHang;
 
-public interface DonHangDAO extends JpaRepository<DonHang, Integer>{
+public interface DonHangDAO extends JpaRepository<DonHang, Long>{
 
 }
