+++ src/main/java/com/msstore/DAO/SanPhamDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.msstore.Entity.SanPham;
 
-public interface SanPhamDAO extends JpaRepository<SanPham, Integer>{
+public interface SanPhamDAO extends JpaRepository<SanPham, Long>{
 
 }
